aws_resources/benchmarking/requirements.txt
# Pinned versions for reproducible builds
# All models (YOLO v11, YOLO-NAS) now use Ultralytics for consistent API

ultralytics>=8.3.0
pillow>=10.0.0
torch>=2.1.0
torchvision>=0.16.0
